remove bugs from watching pages
