+++ src/containers/AddWatching.js
@@ -1,0 +1,14 @@
+//  @flow
+import React from 'react';
+import MarketList from './MarketList';
+
+class AddWatching extends React.Component {
+  render() {
+    console.log('add watching');
+    return (
+      <MarketList adding />
+    );
+  }
+}
+
+export default AddWatching;
